+++ src/main/windows/createMainWindow.ts
@@ -24,7 +24,7 @@ export function createMainWindow(): BrowserWindow {
     titleBarStyle: 'hiddenInset',
     trafficLightPosition: { x: 16, y: 15 },
     ...(process.platform === 'darwin'
-      ? { transparent: true, vibrancy: 'menu', visualEffectState: 'active' }
+      ? { vibrancy: 'menu', backgroundColor: '#00000000' }
       : { backgroundColor: '#ffffff' }),
     ...(process.platform === 'linux' ? { icon } : {}),
     webPreferences: {